requirements.txt
fastapi
uvicorn
python-multipart
motor
PyPDF2
python-dotenv
pydantic